Add DropShardingTableRuleStatementUpdater
